+++ src/parallel/decode.h
@@ -11,11 +11,11 @@
 /*
 ** function to decode a bitstream recursively
 */
-unsigned char getCharFromHuffmanEncodedBitStream(unsigned char buffer[], int* nbytes, int* nbits, Node* node);
+unsigned char getCharFromHuffmanEncodedBitStream(unsigned char buffer[], ull* nbytes, int* nbits, Node* node);
 
 bool getFrequenciesFromEncodedFile(FILE* inputFile, Dictionary* dict);
-ull* getChunkOffsetsFromEncodedFile(FILE* inputFile, int* numChunks);
-ull* getOriginalChunkSizesFromEncodedFile(FILE* inputFile, int* numChunks);
+ull* getChunkOffsetsFromEncodedFile(FILE* inputFile, ull* numChunks);
+ull* getOriginalChunkSizesFromEncodedFile(FILE* inputFile, ull* numChunks);
 
 /*
 ** Function to decode a inputChunk to a outputChunk according to huffmanAlphabet
@@ -25,7 +25,7 @@ bool chunkDecoder(unsigned char inputChunk[], unsigned char outputChunk[], Node*
 /*
 ** Function to decode a input file to an output file according to given alphabet and chunk sizes
 */
-bool fileDecoder(FILE* inputFile, FILE* outputFile, Node* huffmanTree, ull inputChunkOffsets[], ull inputChunkSizes[], int numOfChunks, int num_threads) ;
+bool fileDecoder(FILE* inputFile, FILE* outputFile, Node* huffmanTree, ull inputChunkOffsets[], ull inputChunkSizes[], ull numOfChunks, int num_threads) ;
 
 /*
 ** Function to perform end to end Huffman file decoding
