+++ src/parallel/frequency.h
@@ -7,7 +7,7 @@
 /*
 ** function that counts the chars in a chunk
 */ 
-void countChunk(unsigned char *chunk,int size,Dictionary *d);
+void countChunk(unsigned char *chunk,ull size,Dictionary *d);
 
 /* 
 ** function that reads chunks from readfile, applies processChunk to each chunk, and writes the result in writefile
